backend/requirements.txt
flask==3.0.0
flask-cors==5.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
python-dotenv==1.0.0
psycopg2-binary==2.9.9
google-auth==2.23.4
google-auth-oauthlib==1.1.0
flask-dance==7.0.0
flask-mail==0.9.1

